Move RITS view endpoints into their own file
